fix parsing bug with mutliple equal non closing statements
If several statements are found inside a block that have the same
originalText, the blocks are stored in the doc in the wrong order.
This happens because the parser uses the method string.replace that
always replaces the first instance of a string but the blocks are
in revered order on the stack.